Top 3 Highest-Rated IMDb Movies - Web Page
This project is a simple HTML-based webpage showcasing the top 3 highest-rated movies on IMDb. The webpage includes:

A homepage introducing the concept and listing the movies
Detailed descriptions of each film, including director, cast, and IMDb ratings
An "About Us" section explaining the purpose of the site and the team behind it
Designed for movie enthusiasts, this site provides an easy-to-navigate experience for discovering some of the greatest films of all time.